requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==2.3.6
sqlalchemy==2.0.19
email-validator==2.0.0
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
